% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/launch_vizrd.r
\name{explore_all_data}
\alias{explore_all_data}
\title{Open an interactive browser window to explore
all datasets (data.frame and inheritors) in the global
environment.}
\usage{
explore_all_data()
}
\description{
Open an interactive browser window to explore
all datasets (data.frame and inheritors) in the global
environment.
}
\examples{
explore_all_data()
}

