% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/launch_vizrd.r
\name{explore_data}
\alias{explore_data}
\title{Open an interactive browser window to explore
the dataset through shiny.}
\usage{
explore_data(dfs)
}
\arguments{
\item{dfs}{A vector of dataset names to use.}
}
\description{
Open an interactive browser window to explore
the dataset through shiny.
}
\examples{
explore_data(c("iris","mtcars"))
}

