% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PlotsModel.r
\name{iplotROC}
\alias{iplotROC}
\title{Typical ROC plot, with ggvis hover for cutoff point.}
\usage{
iplotROC(prob, obs)
}
\arguments{
\item{prob}{predicted probability}

\item{obs}{observed labels}
}
\description{
Typical ROC plot, with ggvis hover for cutoff point.
}

