% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotsModel.r
\name{plotCM}
\alias{plotCM}
\title{Bar plot of confusion matrices}
\usage{
plotCM(pred, obs)
}
\arguments{
\item{pred}{predicted labels.}

\item{obs}{observed labels.}
}
\description{
Bar plot of confusion matrices
}

