% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotsModel.r
\name{plotCMProbs}
\alias{plotCMProbs}
\title{Plots the ratios of TP, TN, FP, and FN at each probability cutoff value.}
\usage{
plotCMProbs(prob, obs, posLabel, negLabel, probSeq = seq(0.005, 0.995, by =
  0.005))
}
\arguments{
\item{prob}{predicted probability}

\item{obs}{observed labels}

\item{posLabel}{'positive' label}

\item{negLabel}{'negative' label}

\item{probSeq}{the probability threshold sequence to use}
}
\description{
Plots the ratios of TP, TN, FP, and FN at each probability cutoff value.
}

