% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotsExplore.r
\name{plotDataDist}
\alias{plotDataDist}
\title{Plots the distribution for a given column(s) or for a column by group.}
\usage{
plotDataDist(dat, colNms = NULL, byCol = NULL, violCol = "gray70",
  outAlpha = 0.5, yLogScale = FALSE)
}
\arguments{
\item{dat}{dataset}

\item{colNms}{the column(s) you want to plot}

\item{byCol}{by group column name by which to facet the plots}

\item{violCol}{the colour of the violin}

\item{outAlpha}{the transparency for the boxplot outliers}

\item{yLogScale}{if true, plot the y-axis on a log-10 scale}
}
\description{
Plots the distribution for a given column(s) or for a column by group.
}

