% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotsExplore.r
\name{plotDataHeatmap}
\alias{plotDataHeatmap}
\title{Plots a heatmap representation fo the data.}
\usage{
plotDataHeatmap(dat, colNms, numObs = 75)
}
\arguments{
\item{dat}{dataset}

\item{colNms}{a vector of column names you want to include (all factor or character columns removed)}

\item{numObs}{The top n observations to include in the plot}
}
\description{
Plots a heatmap representation fo the data.
}

