% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotsExplore.r
\name{plotDataPoints}
\alias{plotDataPoints}
\title{Plots raw data points (y-axis = values, x-axis = row #).}
\usage{
plotDataPoints(dat, colNms = names(dat), numObs = NULL, byCol = NULL)
}
\arguments{
\item{dat}{dataset}

\item{colNms}{the column(s) you want to plot}

\item{numObs}{the top n observations to include in the plot}

\item{byCol}{by group column name by which to facet the plots}
}
\description{
Plots raw data points (y-axis = values, x-axis = row #).
}

