% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotsExplore.r
\name{plotDistToPDF}
\alias{plotDistToPDF}
\title{Exports the raw data point plots (plotDataPoints()) to a PDF
document.  Note: this can be a very large document depending on
the number of observations and variables since it'll save the PDF
as vector graphics so each point will be rendered.}
\usage{
plotDistToPDF(dat, colNms = NULL, byCol = NULL, violCol = "gray70",
  outAlpha = 0.5, totPerPage = 9, pdfFile)
}
\arguments{
\item{dat}{dataset}

\item{byCol}{by group column name by which to facet the plots}

\item{violCol}{the colour of the violin}

\item{outAlpha}{the transparency for the boxplot outliers}

\item{totPerPage}{Total number of figures per page (default=9)}

\item{pdfFile}{The file path/name to save to.}

\item{colNm}{the column you want to plot}
}
\description{
Exports the raw data point plots (plotDataPoints()) to a PDF
document.  Note: this can be a very large document depending on
the number of observations and variables since it'll save the PDF
as vector graphics so each point will be rendered.
}

