% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotsModel.r
\name{plotLift}
\alias{plotLift}
\title{Plots the lift for binomial response.}
\usage{
plotLift(prob, obs, posLabel, negLabel)
}
\arguments{
\item{prob}{predicted probability}

\item{obs}{observed labels}

\item{posLabel}{'positive' label}

\item{negLabel}{'negative' label}

\item{probSeq}{the probability threshold sequence to use}
}
\description{
Plots the lift for binomial response.
}

