% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotsModel.r
\name{plotObsPred}
\alias{plotObsPred}
\title{Observed versus predicted values with residuals.}
\usage{
plotObsPred(obs, pred, expandFac = 0.25)
}
\arguments{
\item{obs}{observed values}

\item{pred}{predicted values}

\item{expandFac}{scalar to expand the axes (for kernel to fill correctly)}
}
\description{
Observed versus predicted values with residuals.
}

