% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotsModel.r
\name{plotQQNorm}
\alias{plotQQNorm}
\title{QQ plot for a normal distirbution.}
\usage{
plotQQNorm(obs, pred)
}
\arguments{
\item{obs}{observed values}

\item{pred}{predicted values}
}
\description{
QQ plot for a normal distirbution.
}

