% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotsModel.r
\name{plotResids}
\alias{plotResids}
\title{Plots the standardized residuals versus the predicted values.}
\usage{
plotResids(obs, pred, expandFac = 0.25)
}
\arguments{
\item{obs}{observed values}

\item{pred}{predicted values}

\item{expandFac}{scalar to expand the axes (for kernel to fill correctly)}
}
\description{
Plots the standardized residuals versus the predicted values.
}

