% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotsModel.r
\name{plotVarImpBase}
\alias{plotVarImpBase}
\title{Bar plot of variable importances, passing just a dataframe
of the name, importance, and response (target levels, faceted
if multinomial).  Note: user must sort and set up the data frame
properly depending on the model output.}
\usage{
plotVarImpBase(df, xlabel)
}
\arguments{
\item{df}{data frame}
}
\description{
Bar plot of variable importances, passing just a dataframe
of the name, importance, and response (target levels, faceted
if multinomial).  Note: user must sort and set up the data frame
properly depending on the model output.
}

