% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PlotsExplore.r
\name{saveDefPlot}
\alias{saveDefPlot}
\title{Wrapper for ggsave() with high resolution defaults for suitable
raster-based graphics.}
\usage{
saveDefPlot(filename = "plot.png", units = "in", height = 4, width = 6,
  dpi = 300)
}
\arguments{
\item{filename}{what you want to call it (default: plot.png)}

\item{units}{units of measure (default: inches)}

\item{height}{height of plot in units above (default: 4 in)}

\item{width}{width of plot in units above (default: 6 in)}

\item{dpi}{resolution (default: 300 dpi)}
}
\description{
Wrapper for ggsave() with high resolution defaults for suitable
raster-based graphics.
}

