% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PlotsModel.r
\name{splotCM}
\alias{splotCM}
\title{Bar plot of confusion matrices}
\usage{
splotCM(pred, obs)
}
\arguments{
\item{pred}{predicted labels.}

\item{obs}{observed labels.}
}
\description{
Bar plot of confusion matrices
}

