% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PlotsModel.r
\name{splotCMSeq}
\alias{splotCMSeq}
\title{Bar plot of confusion matrices faceted for different probability thresholds.}
\usage{
splotCMSeq(prob, obs, posLabel, negLabel, probSeq = seq(0.1, 0.9, by = 0.1))
}
\arguments{
\item{prob}{probability values}

\item{obs}{observed labels}

\item{posLabel}{'positive' label}

\item{negLabel}{'negative' label}

\item{probSeq}{the probability threshold sequence to use}
}
\description{
Bar plot of confusion matrices faceted for different probability thresholds.
}

