% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PlotsExplore.r
\name{splotDataDist}
\alias{splotDataDist}
\title{Plots the distribution for a given column(s) or for a column by group.}
\usage{
splotDataDist(dat, colNms = names(dat), byCol = NULL, violCol = "gray70",
  outAlpha = 0.5)
}
\arguments{
\item{dat}{dataset}

\item{colNms}{the column(s) you want to plot}

\item{byCol}{by group column name by which to facet the plots}

\item{violCol}{the colour of the violin}

\item{outAlpha}{the transparency for the boxplot outliers}
}
\description{
Plots the distribution for a given column(s) or for a column by group.
}

