% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PlotsExplore.r
\name{splotDataPoints}
\alias{splotDataPoints}
\title{Plots raw data points (y-axis = values, x-axis = row #).}
\usage{
splotDataPoints(dat, colNms, numObs = NULL, byCol = NULL)
}
\arguments{
\item{dat}{dataset}

\item{colNms}{the column(s) you want to plot}

\item{numObs}{the top n observations to include in the plot}

\item{byCol}{by group column name by which to facet the plots}
}
\description{
Plots raw data points (y-axis = values, x-axis = row #).
}

