% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PlotsExplore.r
\name{splotDistToPDF}
\alias{splotDistToPDF}
\title{Exports the raw data point plots (splotDataPoints()) to a PDF
document.  Note: this can be a very large document depending on
the number of observations and variables since it'll save the PDF
as vector graphics so each point will be rendered.}
\usage{
splotDistToPDF(dat, colNms = NULL, byCol = NULL, baseCol = "gray65",
  totPerPage = 9, pdffile)
}
\arguments{
\item{dat}{dataset}

\item{totPerPage}{Total number of figures per page (default=9)}

\item{pdffile}{The file path/name to save to.}

\item{colNm}{the column you want to plot}

\item{numObs}{The top n observations to include in the plot}
}
\description{
Exports the raw data point plots (splotDataPoints()) to a PDF
document.  Note: this can be a very large document depending on
the number of observations and variables since it'll save the PDF
as vector graphics so each point will be rendered.
}

