% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PlotsModel.r
\name{splotGAMSplines}
\alias{splotGAMSplines}
\title{Plots splines from an mgcv gam model (or gamm, just pass mod$gam).
This has less functionality than the plot.gam() function (no all.terms=T, no rug),
it's just a default using ggplot with facet for visual effects. Since
it doesn't have an all.terms=T option, it will only plot the
spline terms.}
\usage{
splotGAMSplines(mod)
}
\arguments{
\item{mod}{gam mod}
}
\description{
Plots splines from an mgcv gam model (or gamm, just pass mod$gam).
This has less functionality than the plot.gam() function (no all.terms=T, no rug),
it's just a default using ggplot with facet for visual effects. Since
it doesn't have an all.terms=T option, it will only plot the
spline terms.
}

