% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PlotsModel.r
\name{splotLift}
\alias{splotLift}
\title{Plots the lift for binomial response.}
\usage{
splotLift(prob, obs, posLabel, negLabel)
}
\arguments{
\item{prob}{predicted probability}

\item{obs}{observed labels}

\item{posLabel}{'positive' label}

\item{negLabel}{'negative' label}

\item{probSeq}{the probability threshold sequence to use}
}
\description{
Plots the lift for binomial response.
}

