% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PlotsModel.r
\name{splotObsPred}
\alias{splotObsPred}
\title{Observed versus predicted values with residuals.}
\usage{
splotObsPred(obs, pred)
}
\arguments{
\item{obs}{observed values}

\item{pred}{predicted values}
}
\description{
Observed versus predicted values with residuals.
}

