% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PartialPlots.r
\name{splotPartialEffs}
\alias{splotPartialEffs}
\title{Plots partial effects plots for any model type, based on the provided predict function (predFnx).  Partial
effects are calculated by specifying the 'type' argument to be either using either (1) type='median': the median
of numeric covariates and most common level of factor covariates; or (2) type='all': the approach in
randomForest::partialPlot where the full dataset is used to calculate a mean prediction for any given
variable value. Note if using type='all', the method will run painfully slow for large datasets and large
ensembles of trees, but will give a more accurate representation, given the data.}
\usage{
splotPartialEffs(mod, dat, predFnx, colNms, type = "median",
  dWeights = rep(1, dim(dat)[1]), CIOn = T, totPerPage = 9,
  pdffile = NULL)
}
\arguments{
\item{mod}{model object}

\item{dat}{the dataset the model was built on, in order to get mean/median values}

\item{predFnx}{a prediction function with arguments 'function(mod, newdata)' - must return numeric values.
If CIOn=TRUE, column names for returned data.frame should be [colNm]_lower and [colNm]_upper, e.g.,
c('pred', 'pred_low', 'pred_high').}

\item{colNms}{names of the predictors to get the values from}

\item{type}{prediction dataset for each variable value: median=use median/most common value for covariates;
all=use full datasaet for covariates}

\item{CIOn}{whether or not to plot confidence bounds, if so assumes that predFnx passes this in}

\item{totPerPage}{Total number of figures per page (default=9)}

\item{pdffile}{The file path/name to save to.}

\item{colNms}{names of the predictors to get the values from}
}
\description{
Note: in randomForest::partialPlot, they scale classificiation probabilities to the log scale, and set any prob of 0
equal to .Machine$double.eps prior to logging.  Subjective choice of what to assign to a prob of 0 will effect
the shape of the resulting partial effect plot, so results form randomForest::partialPlot should be interpreted
cautiously for datasets where classification probabilities contain a substantial number of estimated 0 values.  The
approach used here is to plot whatever is passed through the predFnx, so if pass probabilities, it'll show the
partial effect on the probability directly.
}
\details{
Here the key is to pass an approriate user-defined predFnx, where the function must return
a data frame: a single column for single-value response (e.g., regression), a single
column for binary classification, or multiple columns for multiclass classification.  If only
certain classes are desired from multiclass classification, just have the predFnx return columns
for those predicted probabilities.
}

