% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PlotsExplore.r
\name{splotPointsToPDF}
\alias{splotPointsToPDF}
\title{Exports the raw data point plots (splotDataPoints()) to a PDF
document.  Note: this can be a very large document depending on
the number of observations and variables since it'll save the PDF
as vector graphics so each point will be rendered.}
\usage{
splotPointsToPDF(dat, colNms = NULL, numObs = NULL, totPerPage = 9,
  pdfFile)
}
\arguments{
\item{dat}{dataset}

\item{numObs}{The top n observations to include in the plot}

\item{totPerPage}{Total number of figures per page (default=9)}

\item{pdfFile}{The file path/name to save to.}

\item{colNm}{the column you want to plot}
}
\description{
Exports the raw data point plots (splotDataPoints()) to a PDF
document.  Note: this can be a very large document depending on
the number of observations and variables since it'll save the PDF
as vector graphics so each point will be rendered.
}

