% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PlotsModel.r
\name{splotQQNorm}
\alias{splotQQNorm}
\title{QQ plot for a normal distirbution.}
\usage{
splotQQNorm(obs, pred)
}
\arguments{
\item{obs}{observed values}

\item{pred}{predicted values}
}
\description{
QQ plot for a normal distirbution.
}

