% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PlotsModel.r
\name{splotTPFPProbs}
\alias{splotTPFPProbs}
\title{Ratio of true positives to false positives.}
\usage{
splotTPFPProbs(prob, obs, posLabel, negLabel, probSeq = seq(0.05, 0.95, by =
  0.05))
}
\arguments{
\item{prob}{predicted probability}

\item{obs}{observed labels}

\item{posLabel}{'positive' label}

\item{negLabel}{'negative' label}

\item{probSeq}{the probability threshold sequence to use}
}
\description{
Ratio of true positives to false positives.
}

