% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PlotsModel.r
\name{splotVarImp}
\alias{splotVarImp}
\title{Bar plot of variable importances.}
\usage{
splotVarImp(name, importance, numVars = 25, responseOrder = NULL,
  xlabel = "Variable Importance")
}
\arguments{
\item{name}{names of variables (sorted by order of plot, high to low)}

\item{importance}{of variables (sorted by order of plot, high to low)}

\item{numVars}{the total number of variables to plot (default=25)}

\item{responseOrder}{the order of responses if multinomial}
}
\value{
plot
}
\description{
Bar plot of variable importances.
}

