% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PlotsExplore.r
\name{splotDataDist}
\alias{splotDataDist}
\title{Plots the distribution for a given column(s) or for a column by group.}
\usage{
splotDataDist(dat, colNms, byCol = NULL, baseCol = "gray65")
}
\arguments{
\item{dat}{dataset}

\item{colNms}{the column(s) you want to plot}

\item{byCol}{by group column name by which to facet the plots}

\item{numObs}{the top n observations to include in the plot}
}
\description{
Plots the distribution for a given column(s) or for a column by group.
}

