% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PlotsExplore.r
\name{splotDataPoints}
\alias{splotDataPoints}
\title{Plots raw data points (y-axis = values, x-axis = row #).}
\usage{
splotDataPoints(dat, colNms, numObs = NULL)
}
\arguments{
\item{dat}{dataset}

\item{numObs}{The top n observations to include in the plot}

\item{colNm}{the column you want to plot}
}
\description{
Plots raw data points (y-axis = values, x-axis = row #).
}

