% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PlotsModel.r
\name{splotObsPred}
\alias{splotObsPred}
\title{Observed versus predicted values with residuals.}
\usage{
splotObsPred(obs, pred, expandFac = 0.25)
}
\arguments{
\item{obs}{observed values}

\item{pred}{predicted values}

\item{expandFac}{scalar to expand the axes (for kernel to fill correctly)}
}
\description{
Observed versus predicted values with residuals.
}

