% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PlotsModel.r
\name{splotResids}
\alias{splotResids}
\title{Plots the standardized residuals versus the predicted values.}
\usage{
splotResids(obs, pred)
}
\arguments{
\item{obs}{observed values}

\item{pred}{predicted values}
}
\description{
Plots the standardized residuals versus the predicted values.
}

